Move flows into flow subpackage.
